+++ packages/gae-js-core/src/middleware/index.ts
@@ -1,5 +1,5 @@
 export { gaeRequestLogger } from "./gae-request-logger";
-export * from "./handle-async";
+export * from "./async-middleware";
 export { requestAsyncStorage } from "./request-async-storage";
 export * from "./request-timeout";
 export { gaeJsApp } from "./gae-js-app";